refactor(cli): split command parsing, execution, and output helpers
Move key/value argument validation plus option and value parsing into focused modules.

Extract command executors and output renderers for blocklists, daemon commands, dashboards, diagnostics, status, and site output.

Keep the CLI facade as the stable API boundary while reducing top-level module size.
