+++ src/cli/execute.rs
@@ -1,69 +1,71 @@
+#[cfg(test)]
 use std::{
-    env, fs,
-    path::Path,
-    sync::{
-        OnceLock,
-        atomic::{AtomicBool, Ordering},
-    },
-    thread,
+    sync::atomic::Ordering,
     time::{Duration, Instant},
 };
 
 use crate::app::App;
-use crate::config::{
-    run_config_doctor, run_config_migration_assistant, validate_automation_trigger_rules,
-};
-use crate::daemon;
-use crate::feature_inventory::{build_feature_inventory_report, export_feature_inventory_report};
+use crate::config::validate_automation_trigger_rules;
 
+#[cfg(test)]
+use crate::cli::StatusComparisonOptions;
 use crate::cli::{
-    AppConfig, AutomationTriggerRuleConfig, AutomationTriggersCommandOutput, BackupOutput,
-    BlocklistCategoryCommandKind, BlocklistCategoryCommandOutput, BlocklistCategorySummaryOutput,
-    BlocklistProfileCommandKind, BlocklistProfileCommandOutput, BlocklistProfileConfig,
-    BlocklistProfileSummaryOutput, BlocklistSiteCommandKind, BreakGlassCommandOutput,
-    CalendarSyncCommandOutput, CliCommand, CommandKind, DaemonConnectionOutput,
-    DaemonStartCommandOutput, DaemonStatusCommandOutput, DaemonStopCommandOutput, DailyGoalConfig,
-    DailyGoalSnapshot, EditSiteResult, ExportOutput, FeatureInventoryOutput, FocusStats,
-    GoalCarryCommandOutput, GoalCommandOutput, HistoryDashboardCardOutput,
-    HistoryDashboardCommandKind, HistoryDashboardCommandOutput, HistoryKpiCardId,
-    InvalidSiteEntryOutput, InvalidSiteInput, MonthlyGoalConfig, OutputMode, PathBuf, ProfileId,
-    ProfileOutput, ProfileView, RecurringScheduleConfig, RestoreOutput, ScheduleCommandOutput,
+    AppConfig, AutomationTriggerRuleConfig, AutomationTriggersCommandOutput,
+    BreakGlassCommandOutput, CliCommand, CommandKind, DailyGoalConfig, DailyGoalSnapshot,
+    FocusStats, GoalCarryCommandOutput, GoalCommandOutput, MonthlyGoalConfig, OutputMode,
+    ProfileId, ProfileOutput, ProfileView, RecurringScheduleConfig, ScheduleCommandOutput,
     ScheduleDelayCommandOutput, SessionMetadataCommandOutput, SessionTemplateCommandKind,
-    SessionTemplateCommandOutput, SessionTemplateSummaryOutput, SiteAddCommandOutput, SiteBlocker,
-    SiteDeleteCommandOutput, SiteEditCommandOutput, SiteEditValue, SiteListCommandOutput,
-    SiteListTarget, StatusComparisonOptions, StatusOutput, StrictCommandOutput, TaskCommandOutput,
-    TaskGoalCommandOutput, TaskGoalOutput, TemporaryAllowlistStatusOutput,
+    SessionTemplateCommandOutput, SessionTemplateSummaryOutput, StrictCommandOutput,
+    TaskCommandOutput, TaskGoalCommandOutput, TaskGoalOutput, TemporaryAllowlistStatusOutput,
     TemporarySiteAddCommandOutput, ThemeCommandOutput, ThemePreset, TimerCommandOutput,
-    TimerStateOutput, UsageSignalsCommandOutput, WeekdayProfileRuleConfig,
-    WeekdayRulesCommandOutput, WeeklyGoalConfig, available_theme_preset_views,
-    build_blocking_preview_command_output, build_diagnostics_command_output,
-    build_schedule_inspection_output, build_status_output_with_comparison, build_task_goal_output,
-    display_input_value, effective_blocked_sites_for_profile, flush_stdout,
-    print_automation_triggers_command_output, print_backup_output,
-    print_blocking_preview_command_output, print_blocklist_category_command_output,
-    print_blocklist_profile_command_output, print_break_glass_command_output,
-    print_calendar_sync_command_output, print_config_doctor_output, print_config_migration_output,
-    print_daemon_start_command_output, print_daemon_status_command_output,
-    print_daemon_stop_command_output, print_diagnostics_command_output, print_export_output,
-    print_feature_inventory_output, print_goal_carry_command_output, print_goal_command_output,
-    print_history_dashboard_command_output, print_json, print_json_compact, print_profile_output,
-    print_restore_output, print_schedule_command_output, print_schedule_delay_command_output,
+    TimerStateOutput, WeekdayProfileRuleConfig, WeekdayRulesCommandOutput, WeeklyGoalConfig,
+    available_theme_preset_views, build_schedule_inspection_output, build_task_goal_output,
+    print_automation_triggers_command_output, print_break_glass_command_output,
+    print_goal_carry_command_output, print_goal_command_output, print_json, print_profile_output,
+    print_schedule_command_output, print_schedule_delay_command_output,
     print_session_metadata_command_output, print_session_template_command_output,
-    print_site_add_command_output, print_site_delete_command_output,
-    print_site_edit_command_output, print_site_list_command_output, print_status_output,
     print_strict_command_output, print_task_goal_command_output,
     print_temporary_site_add_command_output, print_theme_command_output, print_timer_state_output,
-    print_usage_signals_command_output, print_weekday_rules_command_output, profile_id,
-    profile_view, theme_preset_view, timer_phase_id, timer_status_id,
+    print_weekday_rules_command_output, profile_id, profile_view, theme_preset_view,
+    timer_phase_id, timer_status_id,
 };
 
-const CONFIG_FILE_NAME: &str = "config.toml";
-const STATS_FILE_NAME: &str = "stats.toml";
-const WATCH_INTERRUPT_POLL_INTERVAL: Duration = Duration::from_millis(200);
-const USAGE_SIGNAL_SUMMARY_LIMIT: usize = 5;
+mod blocklists;
+mod daemon_commands;
+mod dashboard;
+mod data;
+mod diagnostics;
+mod status;
 
-static WATCH_INTERRUPTED: AtomicBool = AtomicBool::new(false);
-static WATCH_INTERRUPT_HANDLER: OnceLock<Result<(), String>> = OnceLock::new();
+#[cfg(test)]
+pub(super) use blocklists::{
+    apply_blocklist_profile_command, apply_site_add_command, apply_site_delete_command,
+    apply_site_edit_command,
+};
+use blocklists::{
+    execute_blocklist_category_command, execute_blocklist_profile_command,
+    execute_blocklist_sites_command,
+};
+use daemon_commands::{
+    execute_daemon_start_command, execute_daemon_status_command, execute_daemon_stop_command,
+};
+#[cfg(test)]
+pub(super) use dashboard::apply_history_dashboard_command;
+use dashboard::execute_history_dashboard_command;
+#[cfg(test)]
+use data::redact_calendar_sync_error;
+use data::{
+    execute_backup_command, execute_calendar_sync_command, execute_export_command,
+    execute_feature_inventory_command, execute_restore_command, stats_load_options,
+    stats_save_options,
+};
+use diagnostics::{
+    execute_blocking_preview_command, execute_config_doctor_command,
+    execute_config_migrate_command, execute_diagnostics_command, execute_usage_signals_command,
+};
+use status::execute_status_command;
+#[cfg(test)]
+use status::{WATCH_INTERRUPTED, next_watch_deadline, wait_for_next_watch_tick};
 
 pub(super) fn execute_cli_command(cli_command: CliCommand) -> Result<(), String> {
     if let Some(surface_id) = command_usage_surface_id(&cli_command.kind)
@@ -272,62 +274,6 @@ fn execute_next_command(output: OutputMode) -> Result<(), String> {
     emit_timer_command_output("next", &app, output)
 }
 
-fn execute_daemon_start_command(port: Option<u16>, output: OutputMode) -> Result<(), String> {
-    if daemon::is_daemon_child_process() {
-        return daemon::run_foreground(port);
-    }
-
-    let started = daemon::start_background(port)?;
-    let payload = DaemonStartCommandOutput {
-        action: "daemon-start",
-        already_running: started.already_running,
-        daemon: daemon_connection_output(&started.info),
-    };
-    match output {
-        OutputMode::Text => print_daemon_start_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_daemon_status_command(output: OutputMode) -> Result<(), String> {
-    let status = daemon::status()?;
-    let payload = DaemonStatusCommandOutput {
-        action: "daemon-status",
-        running: status.running,
-        daemon: status.info.as_ref().map(daemon_connection_output),
-    };
-    match output {
-        OutputMode::Text => print_daemon_status_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_daemon_stop_command(output: OutputMode) -> Result<(), String> {
-    let stopped = daemon::stop()?;
-    let payload = DaemonStopCommandOutput {
-        action: "daemon-stop",
-        was_running: stopped.was_running,
-        stopped: stopped.stopped,
-        daemon: stopped.info.as_ref().map(daemon_connection_output),
-    };
-    match output {
-        OutputMode::Text => print_daemon_stop_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn daemon_connection_output(connection: &daemon::DaemonConnectionInfo) -> DaemonConnectionOutput {
-    DaemonConnectionOutput {
-        pid: connection.pid,
-        host: connection.host.clone(),
-        port: connection.port,
-        started_at_epoch_secs: connection.started_at_epoch_secs,
-    }
-}
-
 fn execute_task_command(label: String, output: OutputMode) -> Result<(), String> {
     let mut app = App::new();
     app.record_command_usage_for_cli("task");
@@ -439,97 +385,6 @@ fn execute_task_note_command(value: Option<String>, output: OutputMode) -> Resul
     emit_session_metadata_command_output("task-note", updated, &app, output)
 }
 
-fn execute_blocklist_profile_command(
-    command: BlocklistProfileCommandKind,
-    output: OutputMode,
-) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    let payload = apply_blocklist_profile_command(&mut config, command)?;
-    if payload.updated {
-        config
-            .save()
-            .map_err(|error| format!("Failed to save blocklist profile settings: {error}"))?;
-    }
-
-    match output {
-        OutputMode::Text => print_blocklist_profile_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_blocklist_category_command(
-    command: BlocklistCategoryCommandKind,
-    output: OutputMode,
-) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    let payload = apply_blocklist_category_command(&mut config, command)?;
-    if payload.updated {
-        config
-            .save()
-            .map_err(|error| format!("Failed to save blocklist category settings: {error}"))?;
-    }
-    match output {
-        OutputMode::Text => print_blocklist_category_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_blocklist_sites_command(
-    target: SiteListTarget,
-    command: BlocklistSiteCommandKind,
-    output: OutputMode,
-) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    match command {
-        BlocklistSiteCommandKind::List => {
-            let payload = build_site_list_command_output(&config, target, "site-list");
-            match output {
-                OutputMode::Text => print_site_list_command_output(&payload),
-                OutputMode::Json => print_json(&payload)?,
-            }
-        }
-        BlocklistSiteCommandKind::Add { input } => {
-            let payload = apply_site_add_command(&mut config, target, &input)?;
-            if payload.updated {
-                config
-                    .save()
-                    .map_err(|error| format!("Failed to save site changes: {error}"))?;
-            }
-            match output {
-                OutputMode::Text => print_site_add_command_output(&payload),
-                OutputMode::Json => print_json(&payload)?,
-            }
-        }
-        BlocklistSiteCommandKind::Edit { value } => {
-            let payload = apply_site_edit_command(&mut config, target, &value)?;
-            if payload.updated {
-                config
-                    .save()
-                    .map_err(|error| format!("Failed to save site changes: {error}"))?;
-            }
-            match output {
-                OutputMode::Text => print_site_edit_command_output(&payload),
-                OutputMode::Json => print_json(&payload)?,
-            }
-        }
-        BlocklistSiteCommandKind::Delete { site } => {
-            let payload = apply_site_delete_command(&mut config, target, &site)?;
-            if payload.updated {
-                config
-                    .save()
-                    .map_err(|error| format!("Failed to save site changes: {error}"))?;
-            }
-            match output {
-                OutputMode::Text => print_site_delete_command_output(&payload),
-                OutputMode::Json => print_json(&payload)?,
-            }
-        }
-    }
-    Ok(())
-}
-
 fn execute_allowlist_site_add_temporary_command(
     input: String,
     output: OutputMode,
@@ -596,641 +451,6 @@ fn execute_session_template_command(
     Ok(())
 }
 
-fn execute_history_dashboard_command(
-    command: HistoryDashboardCommandKind,
-    output: OutputMode,
-) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    let payload = apply_history_dashboard_command(&mut config, command)?;
-    if payload.updated {
-        config
-            .save()
-            .map_err(|error| format!("Failed to save history dashboard settings: {error}"))?;
-    }
-    match output {
-        OutputMode::Text => print_history_dashboard_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-pub(super) fn apply_history_dashboard_command(
-    config: &mut AppConfig,
-    command: HistoryDashboardCommandKind,
-) -> Result<HistoryDashboardCommandOutput, String> {
-    let (action, updated) = match command {
-        HistoryDashboardCommandKind::Show => ("history-dashboard", false),
-        HistoryDashboardCommandKind::Pin { card } => {
-            if config.history_dashboard.pinned_cards.contains(&card) {
-                ("history-dashboard-pin", false)
-            } else {
-                let insert_at = config
-                    .history_dashboard
-                    .pinned_cards
-                    .iter()
-                    .position(|candidate| {
-                        card_order_index(&config.history_dashboard.card_order, *candidate)
-                            > card_order_index(&config.history_dashboard.card_order, card)
-                    })
-                    .unwrap_or(config.history_dashboard.pinned_cards.len());
-                config
-                    .history_dashboard
-                    .pinned_cards
-                    .insert(insert_at, card);
-                ("history-dashboard-pin", true)
-            }
-        }
-        HistoryDashboardCommandKind::Unpin { card } => {
-            if let Some(index) = config
-                .history_dashboard
-                .pinned_cards
-                .iter()
-                .position(|candidate| *candidate == card)
-            {
-                if config.history_dashboard.pinned_cards.len() <= 1 {
-                    return Err(
-                        "At least one history dashboard card must remain pinned.".to_string()
-                    );
-                }
-                config.history_dashboard.pinned_cards.remove(index);
-                ("history-dashboard-unpin", true)
-            } else {
-                ("history-dashboard-unpin", false)
-            }
-        }
-        HistoryDashboardCommandKind::SetOrder { order } => {
-            if config.history_dashboard.card_order == order {
-                ("history-dashboard-order", false)
-            } else {
-                config.history_dashboard.card_order = order;
-                ("history-dashboard-order", true)
-            }
-        }
-    };
-
-    config
-        .history_dashboard
-        .pinned_cards
-        .sort_by_key(|card| card_order_index(&config.history_dashboard.card_order, *card));
-    Ok(build_history_dashboard_command_output(
-        config, action, updated,
-    ))
-}
-
-fn build_history_dashboard_command_output(
-    config: &AppConfig,
-    action: &'static str,
-    updated: bool,
-) -> HistoryDashboardCommandOutput {
-    let mut cards = config.history_dashboard.pinned_cards.clone();
-    for card in &config.history_dashboard.card_order {
-        if !cards.contains(card) {
-            cards.push(*card);
-        }
-    }
-    HistoryDashboardCommandOutput {
-        action,
-        updated,
-        card_order: config
-            .history_dashboard
-            .card_order
-            .iter()
-            .map(|card| card.id())
-            .collect(),
-        pinned_cards: config
-            .history_dashboard
-            .pinned_cards
-            .iter()
-            .map(|card| card.id())
-            .collect(),
-        cards: cards
-            .into_iter()
-            .map(|card| HistoryDashboardCardOutput {
-                id: card.id(),
-                label: card.label(),
-                pinned: config.history_dashboard.pinned_cards.contains(&card),
-            })
-            .collect(),
-    }
-}
-
-fn card_order_index(order: &[HistoryKpiCardId], card: HistoryKpiCardId) -> usize {
-    order
-        .iter()
-        .position(|candidate| *candidate == card)
-        .unwrap_or(usize::MAX)
-}
-
-pub(super) fn apply_blocklist_profile_command(
-    config: &mut AppConfig,
-    command: BlocklistProfileCommandKind,
-) -> Result<BlocklistProfileCommandOutput, String> {
-    ensure_blocklist_profiles(config);
-
-    let (action, updated) = match command {
-        BlocklistProfileCommandKind::Select { profile } => {
-            handle_select_blocklist_profile(config, profile)?
-        }
-        BlocklistProfileCommandKind::Create { name } => {
-            handle_create_blocklist_profile(config, name)?
-        }
-        BlocklistProfileCommandKind::Rename { name } => {
-            handle_rename_blocklist_profile(config, name)?
-        }
-        BlocklistProfileCommandKind::Delete => handle_delete_blocklist_profile(config)?,
-    };
-
-    if updated {
-        sync_selected_blocklist_profile(config);
-    }
-    Ok(build_blocklist_profile_command_output(
-        config, action, updated,
-    ))
-}
-
-pub(super) fn apply_blocklist_category_command(
-    config: &mut AppConfig,
-    command: BlocklistCategoryCommandKind,
-) -> Result<BlocklistCategoryCommandOutput, String> {
-    ensure_blocklist_profiles(config);
-    let profile_index = selected_blocklist_profile_index(config);
-    ensure_blocklist_categories(&mut config.blocklist_profiles[profile_index]);
-    let profile = &mut config.blocklist_profiles[profile_index];
-
-    let (action, updated) = match command {
-        BlocklistCategoryCommandKind::Select { category } => (
-            "blocklist-category",
-            handle_select_blocklist_category(profile, category)?,
-        ),
-        BlocklistCategoryCommandKind::Create { name } => (
-            "blocklist-category-create",
-            handle_create_blocklist_category(profile, name)?,
-        ),
-        BlocklistCategoryCommandKind::Rename { name } => (
-            "blocklist-category-rename",
-            handle_rename_blocklist_category(profile, name)?,
-        ),
-        BlocklistCategoryCommandKind::Delete => (
-            "blocklist-category-delete",
-            handle_delete_blocklist_category(profile)?,
-        ),
-    };
-
-    sync_profile_site_mirrors(profile);
-    sync_selected_blocklist_profile(config);
-    Ok(build_blocklist_category_command_output(
-        config, action, updated,
-    ))
-}
-
-fn handle_select_blocklist_category(
-    profile: &mut BlocklistProfileConfig,
-    category: Option<String>,
-) -> Result<bool, String> {
-    let Some(category) = category else {
-        return Ok(false);
-    };
-    let index = profile
-        .categories
-        .iter()
-        .position(|candidate| candidate.name.eq_ignore_ascii_case(category.trim()))
-        .ok_or_else(|| format!("Unknown blocklist category `{category}`."))?;
-    let selected = profile.categories[index].name.clone();
-    if profile.selected_category.eq_ignore_ascii_case(&selected) {
-        return Ok(false);
-    }
-    profile.selected_category = selected;
-    Ok(true)
-}
-
-fn handle_create_blocklist_category(
-    profile: &mut BlocklistProfileConfig,
-    name: String,
-) -> Result<bool, String> {
-    let name = name.trim().to_string();
-    if name.is_empty() {
-        return Err("Category name cannot be empty.".to_string());
-    }
-    if profile
-        .categories
-        .iter()
-        .any(|category| category.name.eq_ignore_ascii_case(&name))
-    {
-        return Err(format!("Category `{name}` already exists."));
-    }
-    profile
-        .categories
-        .push(crate::config::BlocklistCategoryConfig {
-            name: name.clone(),
-            sites: Vec::new(),
-            allowlist_sites: Vec::new(),
-        });
-    profile.selected_category = name;
-    Ok(true)
-}
-
-fn handle_rename_blocklist_category(
-    profile: &mut BlocklistProfileConfig,
-    name: String,
-) -> Result<bool, String> {
-    let index = selected_blocklist_category_index(profile);
-    let current = profile.categories[index].name.clone();
-    let name = name.trim().to_string();
-    if name.is_empty() {
-        return Err("Category name cannot be empty.".to_string());
-    }
-    if current.eq_ignore_ascii_case(&name) {
-        return Ok(false);
-    }
-    let duplicate = profile
-        .categories
-        .iter()
-        .enumerate()
-        .any(|(candidate_index, category)| {
-            candidate_index != index && category.name.eq_ignore_ascii_case(&name)
-        });
-    if duplicate {
-        return Err(format!("Category `{name}` already exists."));
-    }
-    profile.categories[index].name = name.clone();
-    profile.selected_category = name;
-    Ok(true)
-}
-
-fn handle_delete_blocklist_category(profile: &mut BlocklistProfileConfig) -> Result<bool, String> {
-    if profile.categories.len() <= 1 {
-        return Err("At least one blocklist category is required.".to_string());
-    }
-    let index = selected_blocklist_category_index(profile);
-    profile.categories.remove(index);
-    let next_index = index.min(profile.categories.len().saturating_sub(1));
-    profile.selected_category = profile.categories[next_index].name.clone();
-    Ok(true)
-}
-
-fn handle_select_blocklist_profile(
-    config: &mut AppConfig,
-    profile: Option<String>,
-) -> Result<(&'static str, bool), String> {
-    let mut updated = false;
-    if let Some(profile) = profile {
-        let index = blocklist_profile_index_by_name(&config.blocklist_profiles, &profile)
-            .ok_or_else(|| format!("Unknown blocklist profile `{profile}`."))?;
-        let selected = config.blocklist_profiles[index].name.clone();
-        if !config
-            .selected_blocklist_profile
-            .eq_ignore_ascii_case(&selected)
-        {
-            config.selected_blocklist_profile = selected;
-            updated = true;
-        }
-    }
-    Ok(("blocklist-profile", updated))
-}
-
-fn handle_create_blocklist_profile(
-    config: &mut AppConfig,
-    name: String,
-) -> Result<(&'static str, bool), String> {
-    let name = name.trim().to_string();
-    if name.is_empty() {
-        return Err("Profile name cannot be empty.".to_string());
-    }
-    if blocklist_profile_index_by_name(&config.blocklist_profiles, &name).is_some() {
-        return Err(format!("Profile `{name}` already exists."));
-    }
-    config.blocklist_profiles.push(BlocklistProfileConfig {
-        name: name.clone(),
-        sites: Vec::new(),
-        allowlist_sites: Vec::new(),
-        categories: vec![crate::config::BlocklistCategoryConfig {
-            name: "General".to_string(),
-            sites: Vec::new(),
-            allowlist_sites: Vec::new(),
-        }],
-        selected_category: "General".to_string(),
-    });
-    config.selected_blocklist_profile = name;
-    Ok(("blocklist-profile-create", true))
-}
-
-fn handle_rename_blocklist_profile(
-    config: &mut AppConfig,
-    name: String,
-) -> Result<(&'static str, bool), String> {
-    let index = selected_blocklist_profile_index(config);
-    let current_name = config.blocklist_profiles[index].name.clone();
-    let name = name.trim().to_string();
-    if name.is_empty() {
-        return Err("Profile name cannot be empty.".to_string());
-    }
-    if current_name == name {
-        return Ok(("blocklist-profile-rename", false));
-    }
-
-    let duplicate =
-        config
-            .blocklist_profiles
-            .iter()
-            .enumerate()
-            .any(|(candidate_index, profile)| {
-                candidate_index != index && profile.name.eq_ignore_ascii_case(&name)
-            });
-    if duplicate {
-        return Err(format!("Profile `{name}` already exists."));
-    }
-
-    config.blocklist_profiles[index].name = name.clone();
-    config.selected_blocklist_profile = name;
-    Ok(("blocklist-profile-rename", true))
-}
-
-fn handle_delete_blocklist_profile(config: &mut AppConfig) -> Result<(&'static str, bool), String> {
-    if config.blocklist_profiles.len() <= 1 {
-        return Err("At least one blocklist profile is required.".to_string());
-    }
-    let index = selected_blocklist_profile_index(config);
-    config.blocklist_profiles.remove(index);
-    let next_index = index.min(config.blocklist_profiles.len().saturating_sub(1));
-    config.selected_blocklist_profile = config.blocklist_profiles[next_index].name.clone();
-    Ok(("blocklist-profile-delete", true))
-}
-
-pub(super) fn apply_site_add_command(
-    config: &mut AppConfig,
-    target: SiteListTarget,
-    input: &str,
-) -> Result<SiteAddCommandOutput, String> {
-    ensure_blocklist_profiles(config);
-    let index = selected_blocklist_profile_index(config);
-    let mut working = SiteBlocker::new();
-    let existing_sites = active_profile_sites(config, index, target).to_vec();
-    for site in existing_sites {
-        working.add_site(site);
-    }
-    let result = working.add_sites_from_input(input);
-    let updated = !result.added.is_empty();
-    *active_profile_sites_mut(config, index, target) = working.sites.clone();
-    sync_profile_site_mirrors(&mut config.blocklist_profiles[index]);
-    if updated {
-        sync_selected_blocklist_profile(config);
-    }
-
-    let active_profile = &config.blocklist_profiles[index];
-    Ok(SiteAddCommandOutput {
-        action: "site-add",
-        updated,
-        profile: active_profile.name.clone(),
-        category: active_profile.selected_category.clone(),
-        target,
-        added: result.added,
-        duplicates: result.duplicates,
-        invalid: invalid_site_entries_output(&result.invalid),
-        sites: active_profile_sites(config, index, target).to_vec(),
-        effective_blocked_sites_count: effective_blocked_sites_for_profile(active_profile).len(),
-    })
-}
-
-pub(super) fn apply_site_edit_command(
-    config: &mut AppConfig,
-    target: SiteListTarget,
-    value: &SiteEditValue,
-) -> Result<SiteEditCommandOutput, String> {
-    ensure_blocklist_profiles(config);
-    let index = selected_blocklist_profile_index(config);
-    let previous = value.previous.trim();
-
-    let mut working = SiteBlocker::new();
-    for site in active_profile_sites(config, index, target) {
-        working.add_site(site.clone());
-    }
-    let edit_index = working
-        .sites
-        .iter()
-        .position(|site| site.eq_ignore_ascii_case(previous))
-        .ok_or_else(|| {
-            format!(
-                "Site `{}` was not found in {}.",
-                value.previous,
-                target.id()
-            )
-        })?;
-    let result = working.edit_site_from_input(edit_index, &value.next);
-    match result {
-        EditSiteResult::Updated { old, new } => {
-            *active_profile_sites_mut(config, index, target) = working.sites.clone();
-            sync_profile_site_mirrors(&mut config.blocklist_profiles[index]);
-            sync_selected_blocklist_profile(config);
-            let active_profile = &config.blocklist_profiles[index];
-            Ok(SiteEditCommandOutput {
-                action: "site-edit",
-                updated: true,
-                profile: active_profile.name.clone(),
-                category: active_profile.selected_category.clone(),
-                target,
-                previous: old,
-                current: new,
-                sites: active_profile_sites(config, index, target).to_vec(),
-                effective_blocked_sites_count: effective_blocked_sites_for_profile(active_profile)
-                    .len(),
-            })
-        }
-        EditSiteResult::Unchanged { hostname } => {
-            let active_profile = &config.blocklist_profiles[index];
-            Ok(SiteEditCommandOutput {
-                action: "site-edit",
-                updated: false,
-                profile: active_profile.name.clone(),
-                category: active_profile.selected_category.clone(),
-                target,
-                previous: hostname.clone(),
-                current: hostname,
-                sites: active_profile_sites(config, index, target).to_vec(),
-                effective_blocked_sites_count: effective_blocked_sites_for_profile(active_profile)
-                    .len(),
-            })
-        }
-        EditSiteResult::Duplicate { hostname } => {
-            Err(format!("`{hostname}` already exists in {}.", target.id()))
-        }
-        EditSiteResult::Invalid(invalid) => Err(format!(
-            "Invalid hostname `{}` ({})",
-            display_input_value(&invalid.input),
-            invalid.reason.message()
-        )),
-        EditSiteResult::MissingSelection => Err("No site selected to edit.".to_string()),
-    }
-}
-
-pub(super) fn apply_site_delete_command(
-    config: &mut AppConfig,
-    target: SiteListTarget,
-    site: &str,
-) -> Result<SiteDeleteCommandOutput, String> {
-    ensure_blocklist_profiles(config);
-    let index = selected_blocklist_profile_index(config);
-    let site = site.trim();
-
-    let mut working = SiteBlocker::new();
-    for current in active_profile_sites(config, index, target) {
-        working.add_site(current.clone());
-    }
-    let delete_index = working
-        .sites
-        .iter()
-        .position(|value| value.eq_ignore_ascii_case(site))
-        .ok_or_else(|| format!("Site `{site}` was not found in {}.", target.id()))?;
-    let removed = working
-        .remove_site(delete_index)
-        .ok_or_else(|| format!("Site `{site}` was not found in {}.", target.id()))?;
-    *active_profile_sites_mut(config, index, target) = working.sites.clone();
-    sync_profile_site_mirrors(&mut config.blocklist_profiles[index]);
-    sync_selected_blocklist_profile(config);
-
-    let active_profile = &config.blocklist_profiles[index];
-    Ok(SiteDeleteCommandOutput {
-        action: "site-delete",
-        updated: true,
-        profile: active_profile.name.clone(),
-        category: active_profile.selected_category.clone(),
-        target,
-        removed,
-        sites: active_profile_sites(config, index, target).to_vec(),
-        effective_blocked_sites_count: effective_blocked_sites_for_profile(active_profile).len(),
-    })
-}
-
-fn ensure_blocklist_profiles(config: &mut AppConfig) {
-    if config.blocklist_profiles.is_empty() {
-        config
-            .blocklist_profiles
-            .push(BlocklistProfileConfig::default());
-        config.selected_blocklist_profile = config.blocklist_profiles[0].name.clone();
-    }
-    let index = selected_blocklist_profile_index(config);
-    ensure_blocklist_categories(&mut config.blocklist_profiles[index]);
-}
-
-fn blocklist_profile_index_by_name(
-    profiles: &[BlocklistProfileConfig],
-    name: &str,
-) -> Option<usize> {
-    profiles
-        .iter()
-        .position(|profile| profile.name.eq_ignore_ascii_case(name.trim()))
-}
-
-fn selected_blocklist_profile_index(config: &AppConfig) -> usize {
-    blocklist_profile_index_by_name(
-        &config.blocklist_profiles,
-        &config.selected_blocklist_profile,
-    )
-    .unwrap_or(0)
-}
-
-fn active_profile_sites(
-    config: &AppConfig,
-    profile_index: usize,
-    target: SiteListTarget,
-) -> &[String] {
-    let profile = &config.blocklist_profiles[profile_index];
-    if profile.categories.is_empty() {
-        return match target {
-            SiteListTarget::Blocklist => &profile.sites,
-            SiteListTarget::Allowlist => &profile.allowlist_sites,
-        };
-    }
-    let category_index = selected_blocklist_category_index(profile);
-    let category = &profile.categories[category_index];
-    match target {
-        SiteListTarget::Blocklist => &category.sites,
-        SiteListTarget::Allowlist => &category.allowlist_sites,
-    }
-}
-
-fn active_profile_sites_mut(
-    config: &mut AppConfig,
-    profile_index: usize,
-    target: SiteListTarget,
-) -> &mut Vec<String> {
-    ensure_blocklist_categories(&mut config.blocklist_profiles[profile_index]);
-    let category_index =
-        selected_blocklist_category_index(&config.blocklist_profiles[profile_index]);
-    let category = &mut config.blocklist_profiles[profile_index].categories[category_index];
-    match target {
-        SiteListTarget::Blocklist => &mut category.sites,
-        SiteListTarget::Allowlist => &mut category.allowlist_sites,
-    }
-}
-
-fn sync_selected_blocklist_profile(config: &mut AppConfig) {
-    ensure_blocklist_profiles(config);
-    let index = selected_blocklist_profile_index(config);
-    config.selected_blocklist_profile = config.blocklist_profiles[index].name.clone();
-}
-
-fn build_blocklist_profile_command_output(
-    config: &AppConfig,
-    action: &'static str,
-    updated: bool,
-) -> BlocklistProfileCommandOutput {
-    let selected_name = config
-        .blocklist_profiles
-        .get(selected_blocklist_profile_index(config))
-        .map(|profile| profile.name.clone())
-        .unwrap_or_else(|| config.selected_blocklist_profile.clone());
-    let profiles = config
-        .blocklist_profiles
-        .iter()
-        .map(|profile| BlocklistProfileSummaryOutput {
-            name: profile.name.clone(),
-            active: profile.name.eq_ignore_ascii_case(&selected_name),
-            blocklist_sites_count: profile.sites.len(),
-            allowlist_sites_count: profile.allowlist_sites.len(),
-            effective_blocked_sites_count: effective_blocked_sites_for_profile(profile).len(),
-        })
-        .collect();
-
-    BlocklistProfileCommandOutput {
-        action,
-        updated,
-        selected_blocklist_profile: selected_name,
-        profiles,
-    }
-}
-
-fn build_blocklist_category_command_output(
-    config: &AppConfig,
-    action: &'static str,
-    updated: bool,
-) -> BlocklistCategoryCommandOutput {
-    let profile_index = selected_blocklist_profile_index(config);
-    let profile = &config.blocklist_profiles[profile_index];
-    let selected_category = if profile.categories.is_empty() {
-        "General".to_string()
-    } else {
-        let index = selected_blocklist_category_index(profile);
-        profile.categories[index].name.clone()
-    };
-    let categories = profile
-        .categories
-        .iter()
-        .map(|category| BlocklistCategorySummaryOutput {
-            name: category.name.clone(),
-            active: category.name.eq_ignore_ascii_case(&selected_category),
-            blocklist_sites_count: category.sites.len(),
-            allowlist_sites_count: category.allowlist_sites.len(),
-        })
-        .collect();
-    BlocklistCategoryCommandOutput {
-        action,
-        updated,
-        selected_blocklist_profile: profile.name.clone(),
-        selected_blocklist_category: selected_category,
-        categories,
-    }
-}
-
 fn build_session_template_command_output(
     app: &App,
     action: &'static str,
@@ -1257,117 +477,6 @@ fn build_session_template_command_output(
         selected_session_template,
         templates,
     }
-}
-
-fn build_site_list_command_output(
-    config: &AppConfig,
-    target: SiteListTarget,
-    action: &'static str,
-) -> SiteListCommandOutput {
-    if config.blocklist_profiles.is_empty() {
-        let fallback = if config.selected_blocklist_profile.trim().is_empty() {
-            "Default".to_string()
-        } else {
-            config.selected_blocklist_profile.clone()
-        };
-        return SiteListCommandOutput {
-            action,
-            profile: fallback,
-            category: "General".to_string(),
-            target,
-            sites: Vec::new(),
-            effective_blocked_sites_count: 0,
-        };
-    }
-    let index = selected_blocklist_profile_index(config);
-    let profile = &config.blocklist_profiles[index];
-    SiteListCommandOutput {
-        action,
-        profile: profile.name.clone(),
-        category: profile.selected_category.clone(),
-        target,
-        sites: active_profile_sites(config, index, target).to_vec(),
-        effective_blocked_sites_count: effective_blocked_sites_for_profile(profile).len(),
-    }
-}
-
-fn invalid_site_entries_output(values: &[InvalidSiteInput]) -> Vec<InvalidSiteEntryOutput> {
-    values
-        .iter()
-        .map(|invalid| InvalidSiteEntryOutput {
-            input: invalid.input.clone(),
-            reason: invalid.reason.message().to_string(),
-        })
-        .collect()
-}
-
-fn ensure_blocklist_categories(profile: &mut BlocklistProfileConfig) {
-    if profile.categories.is_empty() {
-        profile
-            .categories
-            .push(crate::config::BlocklistCategoryConfig {
-                name: "General".to_string(),
-                sites: profile.sites.clone(),
-                allowlist_sites: profile.allowlist_sites.clone(),
-            });
-    }
-    let selected = profile.selected_category.trim().to_string();
-    if selected.is_empty() {
-        profile.selected_category = profile
-            .categories
-            .first()
-            .map(|category| category.name.clone())
-            .unwrap_or_else(|| "General".to_string());
-    } else if let Some(category) = profile
-        .categories
-        .iter()
-        .find(|category| category.name.eq_ignore_ascii_case(&selected))
-    {
-        profile.selected_category = category.name.clone();
-    } else {
-        profile.selected_category = profile
-            .categories
-            .first()
-            .map(|category| category.name.clone())
-            .unwrap_or_else(|| "General".to_string());
-    }
-}
-
-fn selected_blocklist_category_index(profile: &BlocklistProfileConfig) -> usize {
-    profile
-        .categories
-        .iter()
-        .position(|category| {
-            category
-                .name
-                .eq_ignore_ascii_case(&profile.selected_category)
-        })
-        .unwrap_or(0)
-}
-
-fn sync_profile_site_mirrors(profile: &mut BlocklistProfileConfig) {
-    let mut sites: Vec<String> = Vec::new();
-    let mut allowlist_sites: Vec<String> = Vec::new();
-    for category in &profile.categories {
-        for site in &category.sites {
-            if !sites
-                .iter()
-                .any(|existing| existing.eq_ignore_ascii_case(site))
-            {
-                sites.push(site.clone());
-            }
-        }
-        for site in &category.allowlist_sites {
-            if !allowlist_sites
-                .iter()
-                .any(|existing| existing.eq_ignore_ascii_case(site))
-            {
-                allowlist_sites.push(site.clone());
-            }
-        }
-    }
-    profile.sites = sites;
-    profile.allowlist_sites = allowlist_sites;
 }
 
 fn execute_profile_command(profile: Option<ProfileId>, output: OutputMode) -> Result<(), String> {
@@ -1726,682 +835,6 @@ fn execute_break_glass_cancel_command(output: OutputMode) -> Result<(), String> 
     app.record_command_usage_for_cli("break-glass-cancel");
     app.cancel_break_glass_for_cli()?;
     emit_break_glass_command_output("break-glass-cancel", &app, output)
-}
-
-fn execute_config_doctor_command(output: OutputMode) -> Result<(), String> {
-    let payload = run_config_doctor();
-    match output {
-        OutputMode::Text => print_config_doctor_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_config_migrate_command(apply: bool, output: OutputMode) -> Result<(), String> {
-    let payload = run_config_migration_assistant(apply);
-    match output {
-        OutputMode::Text => print_config_migration_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_diagnostics_command(output: OutputMode) -> Result<(), String> {
-    let mut app = App::new();
-    app.record_command_usage_for_cli("diagnostics");
-    let payload = build_diagnostics_command_output(&app.setup_diagnostics);
-
-    match output {
-        OutputMode::Text => print_diagnostics_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_blocking_preview_command(output: OutputMode) -> Result<(), String> {
-    let mut app = App::new();
-    app.record_command_usage_for_cli("blocking-preview");
-    let preview = app.blocking_preview_for_cli()?;
-    let payload = build_blocking_preview_command_output(&preview);
-
-    match output {
-        OutputMode::Text => print_blocking_preview_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_usage_signals_command(output: OutputMode) -> Result<(), String> {
-    let config = AppConfig::load().normalized();
-    let stats = FocusStats::load_with_options(stats_load_options(&config))
-        .map_err(|error| format!("Failed to load stats: {error}"))?;
-    let payload = UsageSignalsCommandOutput {
-        action: "usage-signals",
-        summary: stats.usage_signal_summary(USAGE_SIGNAL_SUMMARY_LIMIT),
-    };
-    match output {
-        OutputMode::Text => print_usage_signals_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_status_command(
-    output: OutputMode,
-    watch_interval_secs: Option<u64>,
-    comparison: StatusComparisonOptions,
-) -> Result<(), String> {
-    if let Some(interval_secs) = watch_interval_secs {
-        return execute_status_watch_command(output, interval_secs, comparison);
-    }
-
-    let payload = load_status_output(&comparison)?;
-    emit_status_output(&payload, output, false)
-}
-
-fn execute_status_watch_command(
-    output: OutputMode,
-    interval_secs: u64,
-    comparison: StatusComparisonOptions,
-) -> Result<(), String> {
-    if interval_secs == 0 {
-        return Err("`--watch` interval must be greater than 0 seconds.".to_string());
-    }
-
-    install_watch_interrupt_handler()?;
-    WATCH_INTERRUPTED.store(false, Ordering::SeqCst);
-    let interval = Duration::from_secs(interval_secs);
-    let mut next_deadline = Instant::now();
-
-    loop {
-        if WATCH_INTERRUPTED.load(Ordering::SeqCst) {
-            break;
-        }
-
-        let payload = load_status_output(&comparison)?;
-        emit_status_output(&payload, output, true)?;
-        flush_stdout()?;
-
-        next_deadline = next_watch_deadline(next_deadline, interval, Instant::now());
-        if wait_for_next_watch_tick(next_deadline) {
-            break;
-        }
-    }
-
-    Ok(())
-}
-
-fn install_watch_interrupt_handler() -> Result<(), String> {
-    WATCH_INTERRUPT_HANDLER
-        .get_or_init(|| unsafe { install_platform_watch_interrupt_handler() })
-        .clone()
-}
-
-fn next_watch_deadline(previous_deadline: Instant, interval: Duration, now: Instant) -> Instant {
-    let mut deadline = previous_deadline + interval;
-    while deadline <= now {
-        deadline += interval;
-    }
-    deadline
-}
-
-fn wait_for_next_watch_tick(deadline: Instant) -> bool {
-    loop {
-        if WATCH_INTERRUPTED.load(Ordering::SeqCst) {
-            return true;
-        }
-
-        let now = Instant::now();
-        if now >= deadline {
-            return false;
-        }
-
-        let sleep_for = deadline
-            .saturating_duration_since(now)
-            .min(WATCH_INTERRUPT_POLL_INTERVAL);
-        thread::sleep(sleep_for);
-    }
-}
-
-#[cfg(unix)]
-unsafe fn install_platform_watch_interrupt_handler() -> Result<(), String> {
-    unsafe extern "C" fn handle_sigint(_signal: i32) {
-        WATCH_INTERRUPTED.store(true, Ordering::SeqCst);
-    }
-
-    unsafe extern "C" {
-        fn signal(signum: i32, handler: unsafe extern "C" fn(i32)) -> unsafe extern "C" fn(i32);
-    }
-
-    const SIGINT: i32 = 2;
-
-    let _previous = unsafe { signal(SIGINT, handle_sigint) };
-    Ok(())
-}
-
-#[cfg(target_os = "windows")]
-unsafe fn install_platform_watch_interrupt_handler() -> Result<(), String> {
-    unsafe extern "system" fn handle_console_ctrl(ctrl_type: u32) -> i32 {
-        const CTRL_C_EVENT: u32 = 0;
-        const CTRL_BREAK_EVENT: u32 = 1;
-        if ctrl_type == CTRL_C_EVENT || ctrl_type == CTRL_BREAK_EVENT {
-            WATCH_INTERRUPTED.store(true, Ordering::SeqCst);
-            return 1;
-        }
-        0
-    }
-
-    unsafe extern "system" {
-        fn SetConsoleCtrlHandler(
-            handler_routine: Option<unsafe extern "system" fn(u32) -> i32>,
-            add: i32,
-        ) -> i32;
-    }
-
-    let installed = unsafe { SetConsoleCtrlHandler(Some(handle_console_ctrl), 1) };
-    if installed == 0 {
-        Err(
-            "Failed to install watch interrupt handler: SetConsoleCtrlHandler returned 0."
-                .to_string(),
-        )
-    } else {
-        Ok(())
-    }
-}
-
-#[cfg(not(any(unix, target_os = "windows")))]
-unsafe fn install_platform_watch_interrupt_handler() -> Result<(), String> {
-    Ok(())
-}
-
-fn load_status_output(comparison: &StatusComparisonOptions) -> Result<StatusOutput, String> {
-    let config = AppConfig::load().normalized();
-    let stats = FocusStats::load_with_options(stats_load_options(&config))
-        .map_err(|error| format!("Failed to load stats: {error}"))?;
-    Ok(build_status_output_with_comparison(
-        &config, &stats, comparison,
-    ))
-}
-
-fn emit_status_output(
-    payload: &StatusOutput,
-    output: OutputMode,
-    watch_mode: bool,
-) -> Result<(), String> {
-    match output {
-        OutputMode::Text => {
-            print_status_output(payload);
-            if watch_mode {
-                println!();
-            }
-        }
-        OutputMode::Json => {
-            if watch_mode {
-                print_json_compact(payload)?;
-            } else {
-                print_json(payload)?;
-            }
-        }
-    }
-    Ok(())
-}
-
-fn execute_export_command(dir: Option<PathBuf>, output: OutputMode) -> Result<(), String> {
-    let config = AppConfig::load().normalized();
-    let stats = FocusStats::load_with_options(stats_load_options(&config))
-        .map_err(|error| format!("Failed to load stats: {error}"))?;
-    let history_kpi_context = build_history_kpi_export_context(&config);
-    let target_dir = match dir {
-        Some(path) => path,
-        None => env::current_dir()
-            .map_err(|error| format!("Failed to determine current directory: {error}"))?,
-    };
-    let exported = stats
-        .export_to_dir_with_context(&target_dir, &history_kpi_context)
-        .map_err(|error| format!("Export failed: {error}"))?;
-
-    let payload = ExportOutput {
-        export_dir: target_dir,
-        json_path: exported.json_path,
-        csv_path: exported.csv_path,
-    };
-    match output {
-        OutputMode::Text => print_export_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_feature_inventory_command(
-    dir: Option<PathBuf>,
-    output: OutputMode,
-) -> Result<(), String> {
-    let target_dir = match dir {
-        Some(path) => path,
-        None => env::current_dir()
-            .map_err(|error| format!("Failed to determine current directory: {error}"))?,
-    };
-
-    let report = build_feature_inventory_report();
-    let exported = export_feature_inventory_report(&target_dir, &report)
-        .map_err(|error| format!("Feature inventory export failed: {error}"))?;
-
-    let payload = FeatureInventoryOutput {
-        export_dir: target_dir,
-        json_path: exported.json_path,
-        markdown_path: exported.markdown_path,
-        total_features: report.summary.total_features,
-        keep_count: report.summary.keep_count,
-        merge_count: report.summary.merge_count,
-        remove_count: report.summary.remove_count,
-    };
-    match output {
-        OutputMode::Text => print_feature_inventory_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn build_history_kpi_export_context(config: &AppConfig) -> crate::stats::HistoryKpiExportContext {
-    let selected_automation = config.profile_automation_for(config.selected_profile);
-    crate::stats::HistoryKpiExportContext {
-        reference_day: chrono::Local::now().date_naive(),
-        daily_goal: DailyGoalSnapshot {
-            minutes: config.daily_goal.minutes,
-            pomodoros: config.daily_goal.pomodoros,
-        },
-        weekly_goal: DailyGoalSnapshot {
-            minutes: config.weekly_goal.minutes,
-            pomodoros: config.weekly_goal.pomodoros,
-        },
-        monthly_goal: DailyGoalSnapshot {
-            minutes: config.monthly_goal.minutes,
-            pomodoros: config.monthly_goal.pomodoros,
-        },
-        carry_over_daily: config.goal_carry_over.daily,
-        carry_over_weekly: config.goal_carry_over.weekly,
-        carry_over_monthly: config.goal_carry_over.monthly,
-        recurring_schedule: selected_automation.recurring_schedule,
-        stats_retention: config.stats_retention,
-        comparison_dimension: crate::stats::ComparisonDimension::TaskLabel,
-        comparison_task_filter: None,
-        comparison_profile_filter: None,
-        comparison_time_of_day_filter: None,
-    }
-}
-
-fn execute_backup_command(dir: Option<PathBuf>, output: OutputMode) -> Result<(), String> {
-    let _config = AppConfig::load().normalized();
-    let backup_dir = match dir {
-        Some(path) => path,
-        None => env::current_dir().map_err(|error| {
-            format!("Backup failed: could not determine current directory: {error}")
-        })?,
-    };
-    fs::create_dir_all(&backup_dir)
-        .map_err(|error| format!("Backup failed: could not create backup directory: {error}"))?;
-
-    let source_config = config_file_path()?;
-    let source_stats = stats_persistence_path()?;
-    ensure_backup_source_file(&source_config, CONFIG_FILE_NAME)?;
-    ensure_backup_source_file(&source_stats, STATS_FILE_NAME)?;
-    let config_backup_path = backup_dir.join(CONFIG_FILE_NAME);
-    let stats_backup_path = backup_dir.join(STATS_FILE_NAME);
-
-    copy_file_with_context(&source_config, &config_backup_path, "backup config.toml")?;
-    copy_file_with_context(&source_stats, &stats_backup_path, "backup stats.toml")?;
-
-    let payload = BackupOutput {
-        backup_dir,
-        config_backup_path,
-        stats_backup_path,
-    };
-    match output {
-        OutputMode::Text => print_backup_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_restore_command(dir: Option<PathBuf>, output: OutputMode) -> Result<(), String> {
-    let _config = AppConfig::load().normalized();
-    let restore_dir = match dir {
-        Some(path) => path,
-        None => env::current_dir().map_err(|error| {
-            format!("Restore failed: could not determine current directory: {error}")
-        })?,
-    };
-    let source_config = restore_dir.join(CONFIG_FILE_NAME);
-    let source_stats = restore_dir.join(STATS_FILE_NAME);
-    ensure_restore_source_file(&source_config, CONFIG_FILE_NAME)?;
-    ensure_restore_source_file(&source_stats, STATS_FILE_NAME)?;
-
-    let config_restored_path = config_file_path()?;
-    let stats_restored_path = stats_persistence_path()?;
-    let staged_config_path = temp_restore_path(&config_restored_path, "staged");
-    let staged_stats_path = temp_restore_path(&stats_restored_path, "staged");
-    copy_file_with_context(
-        &source_config,
-        &staged_config_path,
-        "stage restore config.toml",
-    )?;
-    copy_file_with_context(
-        &source_stats,
-        &staged_stats_path,
-        "stage restore stats.toml",
-    )?;
-
-    let original_config_snapshot = snapshot_existing_file(
-        &config_restored_path,
-        "snapshot existing config.toml for rollback",
-    )?;
-    let original_stats_snapshot = snapshot_existing_file(
-        &stats_restored_path,
-        "snapshot existing stats.toml for rollback",
-    )?;
-
-    replace_file_atomically(
-        &staged_config_path,
-        &config_restored_path,
-        "restore config.toml",
-    )?;
-    if let Err(error) = replace_file_atomically(
-        &staged_stats_path,
-        &stats_restored_path,
-        "restore stats.toml",
-    ) {
-        rollback_restored_file(
-            original_config_snapshot.as_deref(),
-            &config_restored_path,
-            "roll back restored config.toml",
-        );
-        rollback_restored_file(
-            original_stats_snapshot.as_deref(),
-            &stats_restored_path,
-            "roll back restored stats.toml",
-        );
-        let _ = remove_file_if_exists(&staged_stats_path);
-        return Err(error);
-    }
-    if let Some(snapshot) = original_config_snapshot.as_deref() {
-        let _ = remove_file_if_exists(snapshot);
-    }
-    if let Some(snapshot) = original_stats_snapshot.as_deref() {
-        let _ = remove_file_if_exists(snapshot);
-    }
-
-    let payload = RestoreOutput {
-        restore_dir,
-        config_restored_path,
-        stats_restored_path,
-    };
-    match output {
-        OutputMode::Text => print_restore_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_calendar_sync_command(output: OutputMode) -> Result<(), String> {
-    let config = AppConfig::load().normalized();
-    let result = crate::calendar::sync_from_config(&config.calendar_sync, chrono::Local::now())?;
-    let errors = result
-        .source_errors
-        .into_iter()
-        .map(redact_calendar_sync_error)
-        .collect::<Vec<_>>();
-    let payload = CalendarSyncCommandOutput {
-        action: "calendar-sync",
-        synced_at_epoch_secs: result.synced_at_epoch_secs,
-        source_count: result.source_count,
-        windows_count: result.windows.len(),
-        error_count: errors.len(),
-        errors,
-    };
-    match output {
-        OutputMode::Text => print_calendar_sync_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn redact_calendar_sync_error(message: String) -> String {
-    message
-        .split_whitespace()
-        .map(redact_calendar_sync_token)
-        .collect::<Vec<_>>()
-        .join(" ")
-}
-
-fn redact_calendar_sync_token(token: &str) -> String {
-    let core_start = token
-        .char_indices()
-        .find(|(_, ch)| !is_wrapper_punctuation(*ch))
-        .map_or(0, |(index, _)| index);
-    let core_end = token
-        .char_indices()
-        .rev()
-        .find(|(_, ch)| !is_wrapper_punctuation(*ch))
-        .map_or(token.len(), |(index, ch)| index + ch.len_utf8());
-    if core_end <= core_start {
-        return token.to_string();
-    }
-    let prefix = &token[..core_start];
-    let core = &token[core_start..core_end];
-    let suffix = &token[core_end..];
-    let redacted_core = if is_url_like_token(core) {
-        redact_url_like_token(core)
-    } else {
-        core.to_string()
-    };
-    format!("{prefix}{redacted_core}{suffix}")
-}
-
-fn is_url_like_token(token: &str) -> bool {
-    token.contains("://")
-        || token.starts_with("webcal://")
-        || token.starts_with("webcals://")
-        || token.starts_with("http://")
-        || token.starts_with("https://")
-}
-
-fn redact_url_like_token(token: &str) -> String {
-    let without_fragment = token.split_once('#').map_or(token, |(base, _)| base);
-    let without_query = without_fragment
-        .split_once('?')
-        .map_or(without_fragment, |(base, _)| base);
-    let Some((scheme, remainder)) = without_query.split_once("://") else {
-        return without_query.to_string();
-    };
-    let authority = remainder.split('/').next().unwrap_or(remainder);
-    let host = authority
-        .rsplit_once('@')
-        .map_or(authority, |(_, stripped)| stripped);
-    if remainder.contains('/') {
-        format!("{scheme}://{host}/<redacted>")
-    } else {
-        format!("{scheme}://{host}")
-    }
-}
-
-fn is_wrapper_punctuation(ch: char) -> bool {
-    matches!(
-        ch,
-        '(' | ')' | '[' | ']' | '{' | '}' | '<' | '>' | '"' | '\'' | ',' | ';'
-    )
-}
-
-fn ensure_restore_source_file(path: &Path, file_name: &str) -> Result<(), String> {
-    if !path.exists() {
-        return Err(format!(
-            "Restore failed: missing `{file_name}` in `{}`.",
-            path.parent()
-                .map(|parent| parent.display().to_string())
-                .unwrap_or_else(|| ".".to_string())
-        ));
-    }
-    if !path.is_file() {
-        return Err(format!(
-            "Restore failed: `{}` is not a regular file.",
-            path.display()
-        ));
-    }
-    Ok(())
-}
-
-fn ensure_backup_source_file(path: &Path, file_name: &str) -> Result<(), String> {
-    if !path.exists() {
-        return Err(format!(
-            "Backup failed: missing `{file_name}` in `{}`.",
-            path.parent()
-                .map(|parent| parent.display().to_string())
-                .unwrap_or_else(|| ".".to_string())
-        ));
-    }
-    if !path.is_file() {
-        return Err(format!(
-            "Backup failed: `{}` is not a regular file.",
-            path.display()
-        ));
-    }
-    Ok(())
-}
-
-fn snapshot_existing_file(path: &Path, context: &str) -> Result<Option<PathBuf>, String> {
-    if !path.exists() {
-        return Ok(None);
-    }
-    if !path.is_file() {
-        return Err(format!(
-            "Failed to {context}: `{}` is not a regular file.",
-            path.display()
-        ));
-    }
-    let snapshot = temp_restore_path(path, "original");
-    fs::copy(path, &snapshot).map_err(|error| {
-        format!(
-            "Failed to {context}: `{}` -> `{}`: {error}",
-            path.display(),
-            snapshot.display()
-        )
-    })?;
-    Ok(Some(snapshot))
-}
-
-fn temp_restore_path(path: &Path, marker: &str) -> PathBuf {
-    let parent = path.parent().unwrap_or_else(|| Path::new("."));
-    let target_name = path
-        .file_name()
-        .and_then(|name| name.to_str())
-        .unwrap_or("focustime-restore");
-    let pid = std::process::id();
-    parent.join(format!(".{target_name}.{pid}.{marker}.tmp"))
-}
-
-fn replace_file_atomically(
-    staged_path: &Path,
-    destination: &Path,
-    context: &str,
-) -> Result<(), String> {
-    #[cfg(target_os = "windows")]
-    {
-        match fs::rename(staged_path, destination) {
-            Ok(()) => Ok(()),
-            Err(error) if error.kind() == std::io::ErrorKind::AlreadyExists => {
-                fs::remove_file(destination).map_err(|remove_error| {
-                    format!(
-                        "Failed to {context}: could not replace `{}`: {remove_error}",
-                        destination.display()
-                    )
-                })?;
-                fs::rename(staged_path, destination).map_err(|rename_error| {
-                    format!(
-                        "Failed to {context}: `{}` -> `{}`: {rename_error}",
-                        staged_path.display(),
-                        destination.display()
-                    )
-                })
-            }
-            Err(error) => {
-                let _ = remove_file_if_exists(staged_path);
-                Err(format!(
-                    "Failed to {context}: `{}` -> `{}`: {error}",
-                    staged_path.display(),
-                    destination.display()
-                ))
-            }
-        }
-    }
-
-    #[cfg(not(target_os = "windows"))]
-    {
-        fs::rename(staged_path, destination).map_err(|error| {
-            let _ = remove_file_if_exists(staged_path);
-            format!(
-                "Failed to {context}: `{}` -> `{}`: {error}",
-                staged_path.display(),
-                destination.display()
-            )
-        })
-    }
-}
-
-fn remove_file_if_exists(path: &Path) -> Result<(), String> {
-    if !path.exists() {
-        return Ok(());
-    }
-    fs::remove_file(path).map_err(|error| format!("Failed to remove `{}`: {error}", path.display()))
-}
-
-fn rollback_restored_file(snapshot: Option<&Path>, destination: &Path, rollback_context: &str) {
-    if let Some(snapshot) = snapshot {
-        let _ = replace_file_atomically(snapshot, destination, rollback_context);
-    } else {
-        let _ = remove_file_if_exists(destination);
-    }
-}
-
-fn config_file_path() -> Result<PathBuf, String> {
-    crate::config::app_data_path(CONFIG_FILE_NAME).ok_or_else(|| {
-        format!(
-            "could not determine application data path for `{CONFIG_FILE_NAME}` (environment is not configured)"
-        )
-    })
-}
-
-fn stats_persistence_path() -> Result<PathBuf, String> {
-    crate::config::stats_data_path(STATS_FILE_NAME).ok_or_else(|| {
-        format!(
-            "could not determine application data path for `{STATS_FILE_NAME}` (environment is not configured)"
-        )
-    })
-}
-
-fn stats_load_options(_config: &AppConfig) -> crate::stats::StatsLoadOptions {
-    crate::stats::StatsLoadOptions::default()
-}
-
-fn stats_save_options(_config: &AppConfig) -> crate::stats::StatsSaveOptions {
-    crate::stats::StatsSaveOptions::default()
-}
-
-fn copy_file_with_context(source: &Path, destination: &Path, context: &str) -> Result<(), String> {
-    if let Some(parent) = destination.parent() {
-        fs::create_dir_all(parent).map_err(|error| {
-            format!(
-                "Failed to {context}: could not create `{}`: {error}",
-                parent.display()
-            )
-        })?;
-    }
-    fs::copy(source, destination).map_err(|error| {
-        format!(
-            "Failed to {context}: `{}` -> `{}`: {error}",
-            source.display(),
-            destination.display()
-        )
-    })?;
-    Ok(())
 }
 
 fn emit_timer_command_output(
