+++ src/cli/parsing.rs
@@ -1,13 +1,28 @@
-use crate::cli::{
-    AutomationTriggerRuleConfig, BlocklistCategoryCommandKind, BlocklistProfileCommandKind,
-    BlocklistSiteCommandKind, CliAction, CliCommand, CommandKind, ComparisonDimension,
-    DEFAULT_STATUS_COMPARISON_LIMIT, DEFAULT_WATCH_INTERVAL_SECS, DailyGoalConfig,
-    HistoryDashboardCommandKind, HistoryKpiCardId, MonthlyGoalConfig, NaiveDate,
-    OneTimeFocusWindowConfig, OutputMode, ParsedToken, PrimaryCommand, ProfileBucket, ProfileId,
-    RecurringFocusWindowConfig, RecurringScheduleConfig, SessionTemplateCommandKind, SiteEditValue,
-    SiteListTarget, StatusComparisonOptions, ThemePreset, TimeOfDayBucket, USAGE_TEXT,
-    WeekdayProfileRuleConfig, WeeklyGoalConfig,
+mod options;
+mod value;
+
+pub(super) use options::{
+    parse_compare_by_value, parse_compare_limit_value, parse_compare_profile_value,
+    parse_compare_time_of_day_value, parse_daemon_port, parse_daemon_port_option,
+    parse_status_comparison_options, parse_watch_interval_option, parse_watch_interval_secs,
+    require_nonempty_key_value,
 };
+pub(super) use value::{
+    parse_automation_triggers_value, parse_goal_carry_value, parse_goal_value,
+    parse_history_dashboard_order_value, parse_history_kpi_card_id, parse_monthly_goal_value,
+    parse_profile_id, parse_schedule_value, parse_site_edit_value, parse_strict_value,
+    parse_task_goal_value, parse_theme_preset, parse_weekday_rules_value, parse_weekly_goal_value,
+};
+
+use crate::cli::{
+    BlocklistCategoryCommandKind, BlocklistProfileCommandKind, BlocklistSiteCommandKind, CliAction,
+    CliCommand, CommandKind, HistoryDashboardCommandKind, OutputMode, ParsedToken, PrimaryCommand,
+    SessionTemplateCommandKind, SiteListTarget, StatusComparisonOptions, USAGE_TEXT,
+};
+#[cfg(test)]
+use crate::cli::{DEFAULT_STATUS_COMPARISON_LIMIT, HistoryKpiCardId};
+#[cfg(test)]
+use crate::stats::{ComparisonDimension, ProfileBucket, TimeOfDayBucket};
 
 pub(super) fn parse_global_tokens(tokens: &[ParsedToken]) -> Result<(bool, OutputMode), String> {
     let show_help = tokens
@@ -734,358 +749,6 @@ pub(super) fn finalize_cli_action(
     }
 }
 
-pub(super) fn parse_profile_id(value: &str) -> Result<ProfileId, String> {
-    match value.trim().to_ascii_lowercase().as_str() {
-        "basic" | "classic" => Ok(ProfileId::Classic),
-        "standard" | "deep-work" | "deep_work" | "deepwork" => Ok(ProfileId::DeepWork),
-        "advanced" | "custom" => Ok(ProfileId::Custom),
-        _ => Err(invalid_usage(&format!(
-            "Invalid profile `{value}`. Use `basic`, `standard`, or `advanced`."
-        ))),
-    }
-}
-
-pub(super) fn parse_theme_preset(value: &str) -> Result<ThemePreset, String> {
-    match value.trim().to_ascii_lowercase().as_str() {
-        "classic" => Ok(ThemePreset::Classic),
-        "high-contrast" | "high_contrast" | "highcontrast" => Ok(ThemePreset::HighContrast),
-        "deuteranopia-friendly"
-        | "deuteranopia_friendly"
-        | "deuteranopiafriendly"
-        | "colorblind-friendly"
-        | "colorblind_friendly"
-        | "colorblindfriendly" => Ok(ThemePreset::DeuteranopiaFriendly),
-        _ => Err(invalid_usage(&format!(
-            "Invalid theme preset `{value}`. Use `classic`, `high-contrast`, or `deuteranopia-friendly`."
-        ))),
-    }
-}
-
-pub(super) fn parse_history_kpi_card_id(value: &str) -> Result<HistoryKpiCardId, String> {
-    let normalized = value.trim();
-    HistoryKpiCardId::from_id(normalized).ok_or_else(|| {
-        invalid_usage(&format!(
-            "Invalid history dashboard card `{normalized}`. Use one of: session_summary, focus_score, goal_streak, focus_risk, weekly_allocation, last_interruption, stats_growth, retention, comparison_filters."
-        ))
-    })
-}
-
-pub(super) fn parse_history_dashboard_order_value(
-    value: &str,
-) -> Result<Vec<HistoryKpiCardId>, String> {
-    let entries: Vec<&str> = value
-        .split(',')
-        .map(str::trim)
-        .filter(|entry| !entry.is_empty())
-        .collect();
-    if entries.is_empty() {
-        return Err(invalid_usage(
-            "`--history-dashboard-order` requires a comma-separated list of KPI card IDs.",
-        ));
-    }
-    let mut order = Vec::with_capacity(entries.len());
-    for entry in entries {
-        let card = parse_history_kpi_card_id(entry)?;
-        if order.contains(&card) {
-            return Err(invalid_usage(&format!(
-                "Duplicate history dashboard card `{}` in `--history-dashboard-order`.",
-                card.id()
-            )));
-        }
-        order.push(card);
-    }
-
-    let required = HistoryKpiCardId::all();
-    if order.len() != required.len() || required.iter().any(|card| !order.contains(card)) {
-        return Err(invalid_usage(
-            "`--history-dashboard-order` must include every KPI card exactly once.",
-        ));
-    }
-    Ok(order)
-}
-
-pub(super) fn parse_task_goal_value(
-    value: &str,
-) -> Result<(String, Option<DailyGoalConfig>), String> {
-    let trimmed = value.trim();
-    if let Some((label_raw, goal_raw)) = trimmed.rsplit_once(':')
-        && goal_raw.contains(',')
-    {
-        let label =
-            require_nonempty_key_value(label_raw, "Task goal requires a task label before `:`.")?
-                .to_string();
-        let (minutes, pomodoros) = parse_goal_components(goal_raw, "--task-goal")?;
-        return Ok((label, Some(DailyGoalConfig { minutes, pomodoros })));
-    }
-
-    let label = require_nonempty_key_value(
-        trimmed,
-        "`--task-goal` requires `LABEL` or `LABEL:MINUTES,POMODOROS`.",
-    )?
-    .to_string();
-    Ok((label, None))
-}
-
-pub(super) fn parse_goal_value(value: &str) -> Result<DailyGoalConfig, String> {
-    let (minutes, pomodoros) = parse_goal_components(value, "--goal")?;
-    Ok(DailyGoalConfig { minutes, pomodoros })
-}
-
-pub(super) fn parse_weekly_goal_value(value: &str) -> Result<WeeklyGoalConfig, String> {
-    let (minutes, pomodoros) = parse_goal_components(value, "--goal-weekly")?;
-    Ok(WeeklyGoalConfig { minutes, pomodoros })
-}
-
-pub(super) fn parse_monthly_goal_value(value: &str) -> Result<MonthlyGoalConfig, String> {
-    let (minutes, pomodoros) = parse_goal_components(value, "--goal-monthly")?;
-    Ok(MonthlyGoalConfig { minutes, pomodoros })
-}
-
-fn parse_goal_components(value: &str, flag: &str) -> Result<(u64, u32), String> {
-    let trimmed = value.trim();
-    let (minutes_raw, pomodoros_raw) = trimmed.split_once(',').ok_or_else(|| {
-        invalid_usage(&format!(
-            "Invalid goal `{value}`. Use `{flag}=MINUTES,POMODOROS` (for example `{flag}=120,4`)."
-        ))
-    })?;
-    let minutes = minutes_raw.trim().parse::<u64>().map_err(|_| {
-        invalid_usage(&format!(
-            "Invalid goal minutes in `{value}` for `{flag}`. Use a non-negative integer."
-        ))
-    })?;
-    let pomodoros = pomodoros_raw.trim().parse::<u32>().map_err(|_| {
-        invalid_usage(&format!(
-            "Invalid goal pomodoros in `{value}` for `{flag}`. Use a non-negative integer."
-        ))
-    })?;
-    Ok((minutes, pomodoros))
-}
-
-pub(super) fn parse_strict_value(value: &str) -> Result<bool, String> {
-    match value.trim().to_ascii_lowercase().as_str() {
-        "on" => Ok(true),
-        "off" => Ok(false),
-        _ => Err(invalid_usage(&format!(
-            "Invalid strict mode `{value}`. Use `--strict=on` or `--strict=off`."
-        ))),
-    }
-}
-
-pub(super) fn parse_goal_carry_value(value: &str) -> Result<bool, String> {
-    match value.trim().to_ascii_lowercase().as_str() {
-        "on" => Ok(true),
-        "off" => Ok(false),
-        _ => Err(invalid_usage(&format!(
-            "Invalid goal carry-over `{value}`. Use `on` or `off`."
-        ))),
-    }
-}
-
-pub(super) fn parse_site_edit_value(value: &str) -> Result<SiteEditValue, String> {
-    let trimmed = value.trim();
-    let (previous, next) = trimmed.split_once('=').ok_or_else(|| {
-        invalid_usage(&format!(
-            "Invalid site edit `{value}`. Use `OLD=NEW` (for example `--blocklist-site-edit=old.com=new.com`)."
-        ))
-    })?;
-    let previous = previous.trim();
-    let next = next.trim();
-    if previous.is_empty() || next.is_empty() {
-        return Err(invalid_usage(
-            "Site edit values must include both `OLD` and `NEW` hostnames.",
-        ));
-    }
-    Ok(SiteEditValue {
-        previous: previous.to_string(),
-        next: next.to_string(),
-    })
-}
-
-pub(super) fn parse_schedule_value(value: &str) -> Result<RecurringScheduleConfig, String> {
-    let schedule = serde_json::from_str::<RecurringScheduleConfig>(value).map_err(|error| {
-        invalid_usage(&format!(
-            "Invalid schedule JSON payload: {error}. Use `--schedule-set='{{\"windows\":[...],\"exception_dates\":[...],\"one_time_windows\":[...]}}'`."
-        ))
-    })?;
-    validate_schedule_value(&schedule)?;
-    Ok(schedule)
-}
-
-pub(super) fn parse_weekday_rules_value(
-    value: &str,
-) -> Result<Vec<WeekdayProfileRuleConfig>, String> {
-    let rules = serde_json::from_str::<Vec<WeekdayProfileRuleConfig>>(value).map_err(|error| {
-        invalid_usage(&format!(
-            "Invalid weekday-rules JSON payload: {error}. Use `--weekday-rules-set='[{{\"day\":\"mon\",\"profile\":\"standard\",\"blocklist_profile\":\"Work\",\"session_template\":\"Deep Flow\"}}]'`."
-        ))
-    })?;
-    validate_weekday_rules_value(&rules)?;
-    Ok(rules)
-}
-
-pub(super) fn parse_automation_triggers_value(
-    value: &str,
-) -> Result<Vec<AutomationTriggerRuleConfig>, String> {
-    serde_json::from_str::<Vec<AutomationTriggerRuleConfig>>(value).map_err(|error| {
-        invalid_usage(&format!(
-            "Invalid automation-triggers JSON payload: {error}. Use `--automation-triggers-set='[{{\"trigger\":{{\"type\":\"time\",\"days\":[\"mon\"],\"at\":\"09:00\"}},\"action\":{{\"type\":\"start_focus\"}}}}]'`."
-        ))
-    })
-}
-
-fn validate_schedule_value(schedule: &RecurringScheduleConfig) -> Result<(), String> {
-    for (index, window) in schedule.windows.iter().enumerate() {
-        validate_schedule_window(window, index)?;
-    }
-    for (index, date) in schedule.exception_dates.iter().enumerate() {
-        validate_schedule_exception_date(date, index)?;
-    }
-    for (index, window) in schedule.one_time_windows.iter().enumerate() {
-        validate_one_time_schedule_window(window, index)?;
-    }
-    Ok(())
-}
-
-fn validate_schedule_window(
-    window: &RecurringFocusWindowConfig,
-    index: usize,
-) -> Result<(), String> {
-    if window.days.is_empty() {
-        return Err(invalid_usage(&format!(
-            "Invalid schedule window at index {index}: `days` must include at least one weekday."
-        )));
-    }
-    for day in &window.days {
-        if !is_valid_schedule_weekday(day) {
-            return Err(invalid_usage(&format!(
-                "Invalid schedule window at index {index}: unknown weekday `{day}`."
-            )));
-        }
-    }
-
-    let start_minutes = parse_schedule_minutes(&window.start).ok_or_else(|| {
-        invalid_usage(&format!(
-            "Invalid schedule window at index {index}: start `{}` must be HH:MM in 24-hour format.",
-            window.start
-        ))
-    })?;
-    let end_minutes = parse_schedule_minutes(&window.end).ok_or_else(|| {
-        invalid_usage(&format!(
-            "Invalid schedule window at index {index}: end `{}` must be HH:MM in 24-hour format.",
-            window.end
-        ))
-    })?;
-
-    if start_minutes >= end_minutes {
-        return Err(invalid_usage(&format!(
-            "Invalid schedule window at index {index}: start must be earlier than end."
-        )));
-    }
-    Ok(())
-}
-
-fn validate_schedule_exception_date(value: &str, index: usize) -> Result<(), String> {
-    NaiveDate::parse_from_str(value.trim(), "%Y-%m-%d").map_err(|_| {
-        invalid_usage(&format!(
-            "Invalid exception date at index {index}: `{value}` must be YYYY-MM-DD."
-        ))
-    })?;
-    Ok(())
-}
-
-fn validate_one_time_schedule_window(
-    window: &OneTimeFocusWindowConfig,
-    index: usize,
-) -> Result<(), String> {
-    NaiveDate::parse_from_str(window.date.trim(), "%Y-%m-%d").map_err(|_| {
-        invalid_usage(&format!(
-            "Invalid one-time window at index {index}: date `{}` must be YYYY-MM-DD.",
-            window.date
-        ))
-    })?;
-
-    let start_minutes = parse_schedule_minutes(&window.start).ok_or_else(|| {
-        invalid_usage(&format!(
-            "Invalid one-time window at index {index}: start `{}` must be HH:MM in 24-hour format.",
-            window.start
-        ))
-    })?;
-    let end_minutes = parse_schedule_minutes(&window.end).ok_or_else(|| {
-        invalid_usage(&format!(
-            "Invalid one-time window at index {index}: end `{}` must be HH:MM in 24-hour format.",
-            window.end
-        ))
-    })?;
-
-    if start_minutes >= end_minutes {
-        return Err(invalid_usage(&format!(
-            "Invalid one-time window at index {index}: start must be earlier than end."
-        )));
-    }
-
-    Ok(())
-}
-
-fn validate_weekday_rules_value(rules: &[WeekdayProfileRuleConfig]) -> Result<(), String> {
-    for (index, rule) in rules.iter().enumerate() {
-        if !is_valid_schedule_weekday(&rule.day) {
-            return Err(invalid_usage(&format!(
-                "Invalid weekday rule at index {index}: unknown day `{}`.",
-                rule.day
-            )));
-        }
-        if rule.blocklist_profile.trim().is_empty() {
-            return Err(invalid_usage(&format!(
-                "Invalid weekday rule at index {index}: `blocklist_profile` cannot be empty."
-            )));
-        }
-        if let Some(template) = rule.session_template.as_deref()
-            && template.trim().is_empty()
-        {
-            return Err(invalid_usage(&format!(
-                "Invalid weekday rule at index {index}: `session_template` cannot be empty when provided."
-            )));
-        }
-    }
-    Ok(())
-}
-
-fn is_valid_schedule_weekday(value: &str) -> bool {
-    matches!(
-        value.trim().to_ascii_lowercase().as_str(),
-        "mon"
-            | "monday"
-            | "tue"
-            | "tues"
-            | "tuesday"
-            | "wed"
-            | "wednesday"
-            | "thu"
-            | "thurs"
-            | "thursday"
-            | "fri"
-            | "friday"
-            | "sat"
-            | "saturday"
-            | "sun"
-            | "sunday"
-    )
-}
-
-fn parse_schedule_minutes(value: &str) -> Option<u16> {
-    let trimmed = value.trim();
-    let (hours, minutes) = trimmed.split_once(':')?;
-    if hours.len() != 2 || minutes.len() != 2 {
-        return None;
-    }
-    let hours = hours.parse::<u16>().ok()?;
-    let minutes = minutes.parse::<u16>().ok()?;
-    if hours > 23 || minutes > 59 {
-        return None;
-    }
-    Some(hours * 60 + minutes)
-}
-
 fn set_primary_command(
     primary: &mut Option<PrimaryCommand>,
     next: PrimaryCommand,
@@ -1176,211 +839,6 @@ fn primary_name(command: &PrimaryCommand) -> &'static str {
 
 pub(super) fn invalid_usage(message: &str) -> String {
     format!("{message}\n\n{USAGE_TEXT}")
-}
-
-pub(super) fn parse_watch_interval_option(tokens: &[ParsedToken]) -> Result<Option<u64>, String> {
-    let mut interval: Option<u64> = None;
-    for token in tokens {
-        if let ParsedToken::Watch(value) = token {
-            if interval.is_some() {
-                return Err(invalid_usage("`--watch` can only be specified once."));
-            }
-            interval = Some(value.unwrap_or(DEFAULT_WATCH_INTERVAL_SECS));
-        }
-    }
-    Ok(interval)
-}
-
-pub(super) fn parse_daemon_port_option(tokens: &[ParsedToken]) -> Result<Option<u16>, String> {
-    let mut port: Option<u16> = None;
-    for token in tokens {
-        if let ParsedToken::DaemonPort(value) = token {
-            if port.is_some() {
-                return Err(invalid_usage("`--daemon-port` can only be specified once."));
-            }
-            port = Some(*value);
-        }
-    }
-    Ok(port)
-}
-
-#[derive(Default)]
-struct StatusComparisonParseFlags {
-    has_dimension: bool,
-    has_task: bool,
-    has_profile: bool,
-    has_time_of_day: bool,
-    has_limit: bool,
-    has_any_option: bool,
-}
-
-fn apply_status_comparison_token(
-    token: &ParsedToken,
-    options: &mut StatusComparisonOptions,
-    flags: &mut StatusComparisonParseFlags,
-) -> Result<(), String> {
-    match token {
-        ParsedToken::CompareBy(dimension) => {
-            flags.has_any_option = true;
-            mark_compare_option_seen(
-                &mut flags.has_dimension,
-                "`--compare-by` can only be specified once.",
-            )?;
-            options.dimension = *dimension;
-        }
-        ParsedToken::CompareTask(task) => {
-            flags.has_any_option = true;
-            mark_compare_option_seen(
-                &mut flags.has_task,
-                "`--compare-task` can only be specified once.",
-            )?;
-            options.task_label = task.clone();
-        }
-        ParsedToken::CompareProfile(profile) => {
-            flags.has_any_option = true;
-            mark_compare_option_seen(
-                &mut flags.has_profile,
-                "`--compare-profile` can only be specified once.",
-            )?;
-            options.profile = *profile;
-        }
-        ParsedToken::CompareTimeOfDay(bucket) => {
-            flags.has_any_option = true;
-            mark_compare_option_seen(
-                &mut flags.has_time_of_day,
-                "`--compare-time` can only be specified once.",
-            )?;
-            options.time_of_day = *bucket;
-        }
-        ParsedToken::CompareLimit(limit) => {
-            flags.has_any_option = true;
-            mark_compare_option_seen(
-                &mut flags.has_limit,
-                "`--compare-limit` can only be specified once.",
-            )?;
-            options.limit = *limit;
-        }
-        _ => {}
-    }
-    Ok(())
-}
-
-fn mark_compare_option_seen(flag: &mut bool, message: &str) -> Result<(), String> {
-    if *flag {
-        return Err(invalid_usage(message));
-    }
-    *flag = true;
-    Ok(())
-}
-
-pub(super) fn parse_status_comparison_options(
-    tokens: &[ParsedToken],
-) -> Result<(StatusComparisonOptions, bool), String> {
-    let mut options = StatusComparisonOptions::default();
-    let mut flags = StatusComparisonParseFlags::default();
-
-    for token in tokens {
-        apply_status_comparison_token(token, &mut options, &mut flags)?;
-    }
-
-    if options.limit == 0 {
-        return Err(invalid_usage(
-            "`--compare-limit` requires a positive whole number.",
-        ));
-    }
-    if !flags.has_limit {
-        options.limit = DEFAULT_STATUS_COMPARISON_LIMIT;
-    }
-    Ok((options, flags.has_any_option))
-}
-
-pub(super) fn parse_compare_by_value(value: &str) -> Result<ComparisonDimension, String> {
-    match value.trim().to_ascii_lowercase().as_str() {
-        "task" | "task-label" | "task_label" => Ok(ComparisonDimension::TaskLabel),
-        "profile" => Ok(ComparisonDimension::Profile),
-        "time-of-day" | "time_of_day" | "timeofday" | "time" => Ok(ComparisonDimension::TimeOfDay),
-        _ => Err(invalid_usage(&format!(
-            "Invalid compare dimension `{value}`. Use `task`, `profile`, or `time-of-day`."
-        ))),
-    }
-}
-
-pub(super) fn parse_compare_profile_value(value: &str) -> Result<Option<ProfileBucket>, String> {
-    match value.trim().to_ascii_lowercase().as_str() {
-        "all" => Ok(None),
-        "basic" | "classic" => Ok(Some(ProfileBucket::Classic)),
-        "standard" | "deep-work" | "deep_work" | "deepwork" => Ok(Some(ProfileBucket::DeepWork)),
-        "advanced" | "custom" => Ok(Some(ProfileBucket::Custom)),
-        "unknown" => Ok(Some(ProfileBucket::Unknown)),
-        _ => Err(invalid_usage(&format!(
-            "Invalid compare profile `{value}`. Use `basic`, `standard`, `advanced`, `unknown`, or `all`."
-        ))),
-    }
-}
-
-pub(super) fn parse_compare_time_of_day_value(
-    value: &str,
-) -> Result<Option<TimeOfDayBucket>, String> {
-    match value.trim().to_ascii_lowercase().as_str() {
-        "all" => Ok(None),
-        "morning" => Ok(Some(TimeOfDayBucket::Morning)),
-        "afternoon" => Ok(Some(TimeOfDayBucket::Afternoon)),
-        "evening" => Ok(Some(TimeOfDayBucket::Evening)),
-        "night" => Ok(Some(TimeOfDayBucket::Night)),
-        "unknown" => Ok(Some(TimeOfDayBucket::Unknown)),
-        _ => Err(invalid_usage(&format!(
-            "Invalid compare time bucket `{value}`. Use `morning`, `afternoon`, `evening`, `night`, `unknown`, or `all`."
-        ))),
-    }
-}
-
-pub(super) fn parse_compare_limit_value(value: &str) -> Result<usize, String> {
-    let trimmed = value.trim();
-    let limit = trimmed
-        .parse::<usize>()
-        .map_err(|_| invalid_usage("`--compare-limit` requires a positive whole number."))?;
-    if limit == 0 {
-        return Err(invalid_usage(
-            "`--compare-limit` requires a positive whole number.",
-        ));
-    }
-    Ok(limit)
-}
-
-pub(super) fn parse_watch_interval_secs(value: &str) -> Result<u64, String> {
-    let trimmed = value.trim();
-    let secs = trimmed
-        .parse::<u64>()
-        .map_err(|_| invalid_usage("`--watch` requires a positive whole number of seconds."))?;
-    if secs == 0 {
-        return Err(invalid_usage(
-            "`--watch` requires a positive whole number of seconds.",
-        ));
-    }
-    Ok(secs)
-}
-
-pub(super) fn parse_daemon_port(value: &str) -> Result<u16, String> {
-    let trimmed = value.trim();
-    let port = trimmed
-        .parse::<u16>()
-        .map_err(|_| invalid_usage("`--daemon-port` requires a port between 1 and 65535."))?;
-    if port == 0 {
-        return Err(invalid_usage(
-            "`--daemon-port` requires a port between 1 and 65535.",
-        ));
-    }
-    Ok(port)
-}
-
-pub(super) fn require_nonempty_key_value<'a>(
-    value: &'a str,
-    message: &str,
-) -> Result<&'a str, String> {
-    if value.trim().is_empty() {
-        return Err(invalid_usage(message));
-    }
-    Ok(value)
 }
 
 #[cfg(test)]
